Optionally manage postgresqld service
If you use PostgreSQL in a cluster environment like Pacemaker,
Puppet should not manage the service.
Additionally, PostgreSQL should not be started on boot in such
an environment.
It is still possible to manage pg_hba rules etc.
The reload manifest remains unchanged, a reload is usually
safe in a cluster environment.